Gjør om generelle tekster i Produktside Dagpenger for kalkulator-tekst
